Handle a recursive failure during initialization
Since cargo-wasi recursively executes itself if we are the shim on an
original installation we need to be careful that only the top-level
process attempts to move itself to the final location rather than all
subsequent processes attempting to do so. That means a child process may
succeed but the parent process won't, causing a failure.